+++ Addons/Destroy/Runtime/Extensions/Ent.Destroy.cs
@@ -10,11 +10,11 @@ namespace ME.BECS {
 
         public static void DestroyEndTick(this in Ent ent) {
 
-            ent.Set(new DestroyWithTicks() { ticks = 0, });
+            ent.Set(new DestroyWithTicks() { ticks = 0UL, });
             
         }
 
-        public static void Destroy(this in Ent ent, int ticks) {
+        public static void Destroy(this in Ent ent, ulong ticks) {
 
             ent.Set(new DestroyWithTicks() { ticks = ticks, });
